backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.0
httpx==0.25.2

# AI/ML dependencies
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Production dependencies
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1
fastapi-limiter==0.1.5
redis==5.0.1
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==1.39.1
python-json-logger==2.0.7

# Development dependencies (not installed in production)
pip-tools==7.3.0
black==23.11.0
isort==5.12.0
mypy==1.7.0